haber_site/requirements.txt
django==5.1.4
pymongo==4.6.1
requests==2.31.0
feedparser==6.0.10
python-decouple==3.8
